gallica/requirements.txt
pyaml
tqdm
beautifulsoup4
requests